+++ Client/IRequest.cs
@@ -70,6 +70,10 @@ namespace Pathoschild.Http.Client
         /// <param name="parameter">The authentication header value (e.g. the bearer token).</param>
         IRequest WithAuthentication(string scheme, string parameter);
 
+        /// <summary>Set whether HTTP error responses (e.g. HTTP 404) should be raised as exceptions for this request.</summary>
+        /// <param name="enabled">Whether to raise HTTP errors as exceptions.</param>
+        IRequest WithHttpErrorAsException(bool enabled);
+
         /// <summary>Set the request coordinator for this request.</summary>
         /// <param name="requestCoordinator">The request coordinator (or null to use the default behaviour).</param>
         IRequest WithRequestCoordinator(IRequestCoordinator requestCoordinator);