+++ RELEASE-NOTES.md
@@ -9,6 +9,7 @@ Upcoming release.
   * Added fluent method to override the default `User-Agent` header.
   * Added support for `IWebProxy`.
   * Added support for cancellation tokens.
+  * Added support for disabling HTTP-errors-as-exceptions per-request or per-client.
 * Breaking changes:
   * Replaced `response.AsList<T>` with `reponse.AsArray<T>`.
   * Removed `JsonNetFormatter` (deprecated since 2.1).